DB data types error corrections.

+++ src/sistemacontable/IngresoPar.java
@@ -413,8 +413,8 @@ public class IngresoPar extends javax.swing.JPanel {
         double sum1 = 0, sum2 = 0;
 
         for (String[] value : lista) {
-            sum1 += Integer.parseInt(value[3]);
-            sum2 += Integer.parseInt(value[4]);
+            sum1 += Double.parseDouble(value[3]);
+            sum2 += Double.parseDouble(value[4]);
         }
 
         return new double[]{sum1, sum2};
